requirements.txt
fastapi~=0.115.11
python-dotenv~=1.0.1
sqlmodel~=0.0.23